fix: PacketizationMode should be *uint8
Related: https://github.com/jiyeyuran/mediasoup-go/issues/24

Like `RtpCodecSpecificParameters.ProfileId`,
there is `h264.RtpParameter.PacketizationMode`.

This value is optional, so it can be in 3 states:
* Undefined.
* Defined, with value 0.
* Defined, with value 1.

So it's more correct that the variable is made optional, as a pointer.

This adds a new function `Uint8()` to make assignments easy.

Note that starting from Go 1.18 with generics, all util pointer
functions can be replaced with a single one like this:

```
func Ptr[T any](v T) *T {
	return &v
}
```

But mediasoup-go is using Go 1.13, so for now the util functions are
needed.