fix: goroutine leaks in tests
Several tests create mediasoup workers and don't close them after
finishing, which leaves leaked goroutines that are blocking PR #31.

This change provides a fix for all those tests. Mostly it just clsoes
resources after using them.

The biggest change is operation of CreateRouter() in helper_test.go:

* If CreateRouter(worker) is called, it just uses it.
* If CreateRouter() is called (without worker), a temporary one was
created and never closed; this change introduces the assumption that the
temporary worker can be closed whenever the router is closed, so it uses
`router.OnClose` to hook on the router closing and closes the worker.

This doesn't change the behavior of CreateRouter, so it is transpaent
for users of this function.

Another change is that the `sctp.Association` in sctp_test.go is also
saved as part of the Suite, because it needs to be explicitly closed
when the test suite finishes.

sctp_test.go used wrongly named variable "stcpStream", now renamed to
"sctpStream" to go with the new "sctpAssociation".